changed the name of the controllers and Jsps to me more usable and understandable, and moved them around for better structure, and solve issues
