+++ pom.xml
@@ -32,6 +32,12 @@
             <version>2.3.3</version>
             <scope>provided</scope>
         </dependency>
+        
+        <dependency>
+    <groupId>javax.servlet.jsp.jstl</groupId>
+    <artifactId>javax.servlet.jsp.jstl-api</artifactId>
+    <version>1.2.1</version>
+</dependency>
 
         <!-- JSTL -->
         <dependency>
