+++ src/main/java/com/javapos/controller/admin/AdminDashboardController.java
@@ -1,4 +1,4 @@
-package com.javapos.controller;
+package com.javapos.controller.admin;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
