+++ src/main/java/com/javapos/filter/AuthenticationFilter.java
@@ -1,12 +1,7 @@
 package com.javapos.filter;
 
 import java.io.IOException;
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
+import javax.servlet.*;
 import javax.servlet.annotation.WebFilter;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -22,12 +17,16 @@ import javax.servlet.http.HttpSession;
     "/Pages/Profile/*"
 })
 public class AuthenticationFilter implements Filter {
+    
+    @Override
     public void init(FilterConfig filterConfig) throws ServletException {
         System.out.println("AuthenticationFilter initialized");
     }
 
+    @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
             throws IOException, ServletException {
+        
         HttpServletRequest httpRequest = (HttpServletRequest) request;
         HttpServletResponse httpResponse = (HttpServletResponse) response;
         HttpSession session = httpRequest.getSession(false);
@@ -43,6 +42,8 @@ public class AuthenticationFilter implements Filter {
         }
     }
 
+    @Override
     public void destroy() {
+        // Cleanup if needed
     }
 }
