+++ src/main/webapp/Pages/Common/footer.jsp
@@ -2,8 +2,18 @@
 
     </div> <!-- Close container from header -->
 
-    <footer class="text-center text-muted mt-5 py-3 border-top">
-        &copy; <%= java.time.Year.now() %> JavaPOS. All rights reserved.
+    <footer class="main-footer">
+        <div class="footer-content">
+            <div class="footer-info">
+                <p>&copy; 2024 Restaurant POS. All rights reserved.</p>
+                <p>Version 1.0.0</p>
+            </div>
+            <div class="footer-links">
+                <a href="${pageContext.request.contextPath}/about">About</a>
+                <a href="${pageContext.request.contextPath}/privacy">Privacy Policy</a>
+                <a href="${pageContext.request.contextPath}/terms">Terms of Service</a>
+            </div>
+        </div>
     </footer>
 
     <script src="<%= request.getContextPath() %>/js/main.js"></script>
