+++ src/main/webapp/Pages/Common/header.jsp
@@ -1,57 +1,26 @@
-<%@ page import="com.javapos.model.User" %>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%
-    User user = (User) session.getAttribute("user");
-    if (user == null) {
-        response.sendRedirect(request.getContextPath() + "/Pages/auth/login.jsp");
-        return;
+    String currentPage = request.getParameter("currentPage");
+    if (currentPage == null) {
+        currentPage = "";
     }
-
-    // Determine dashboard URL based on user role
-    String dashboardUrl = request.getContextPath() + "/Pages/Dashboard/dashboard.jsp";
-    switch (user.getRole().toLowerCase()) {
-        case "admin":
-            dashboardUrl = request.getContextPath() + "/Pages/Dashboard/admin-dashboard.jsp";
-            break;
-        case "cashier":
-            dashboardUrl = request.getContextPath() + "/Pages/Dashboard/cashier-dashboard.jsp";
-            break;
-        case "waiter":
-            dashboardUrl = request.getContextPath() + "/Pages/Dashboard/waiter-dashboard.jsp";
-            break;
-    }
-
-    // Safe fallback for currentPage
-    String currentPage = (String) request.getAttribute("currentPage");
-    if (currentPage == null) currentPage = "";
 %>
 
-<!-- Header Bar -->
-<div class="header-bar">
-    <header>
-        <h1>Point of Sale</h1>
-        <h2>GrandLine</h2>
-    </header>
-</div>
+<header class="main-header">
+    <div class="header-content">
+        <div class="logo">
+            <h1>Restaurant POS</h1>
+        </div>
+        
+        <div class="user-info">
+            <c:if test="${not empty sessionScope.user}">
+                <span class="welcome">Welcome, ${sessionScope.user.fullName}</span>
+                <a href="${pageContext.request.contextPath}/logout" class="btn btn-logout">Logout</a>
+            </c:if>
+        </div>
+    </div>
 
-<!-- Navigation Bar -->
-<div class="nav-bar">
-    <nav>
-        <ul>
-            <% if ("admin".equalsIgnoreCase(user.getRole())) { %>
-                <li><a href="<%= request.getContextPath() %>/Pages/Admin/home.jsp"
-                       class="<%= "Home".equals(currentPage) ? "active" : "" %>">Home</a></li>
-            <% } %>
-
-            <li><a href="<%= dashboardUrl %>"
-                   class="<%= "Dashboard".equals(currentPage) ? "active" : "" %>">Dashboard</a></li>
-
-            <li><a href="<%= request.getContextPath() %>/Pages/Menu/menu-list.jsp"
-                   class="<%= "Menu".equals(currentPage) ? "active" : "" %>">Menu</a></li>
-
-            <li><a href="<%= request.getContextPath() %>/Pages/Profile/user-profile.jsp"
-                   class="<%= "Profile".equals(currentPage) ? "active" : "" %>">View Profile</a></li>
-
-            <li><a href="<%= request.getContextPath() %>/logout">Logout</a></li>
-        </ul>
-    </nav>
-</div>
+    <!-- Include external navigation JSP -->
+    <jsp:include page="/Pages/Common/navigation-bar.jsp" />
+</header>
