+++ src/main/webapp/css/style.css
@@ -1,0 +1,206 @@
+@charset "UTF-8";
+/* Reset */
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+.profile-box {
+    width: 400px;
+    margin: 0 auto; /* Centers horizontally */
+    padding: 20px;
+    background-color: white;
+    border: 1px solid #ccc;
+    border-radius: 12px;
+    text-align: left;
+}
+
+
+/* Body & Typography */
+body {
+    font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+    background-color: #f0f8ff; /* Alice Blue */
+    color: #333;
+}
+
+.container {
+    max-width: 1200px;
+    margin: auto;
+    padding: 20px;
+}
+
+/* Header and Footer */
+header, footer {
+    background-color: #4996d1;
+    color: white;
+    padding: 15px 20px;
+    text-align: center;
+}
+
+/* Center the Navbar */
+.nav-bar {
+    display: flex;
+    justify-content: center; /* Centers the nav bar */
+    align-items: center;
+    background-color: #4996d1;
+    padding: 10px 0;
+}
+
+.nav-list {
+    display: flex;
+    list-style: none;
+    margin: 0;
+    padding: 0;
+}
+
+.nav-item {
+    margin: 0 20px;
+    position: relative;
+}
+
+.nav-item a {
+    color: #fff;
+    text-decoration: none;
+    font-size: 16px;
+    padding: 10px 15px;
+    border-radius: 5px;
+    transition: background-color 0.3s ease;
+}
+
+.nav-item a:hover,
+.nav-item.active a {
+    background-color: #ffe87c; /* Hover and active state color */
+    color: #556b76;
+}
+
+.nav-item a:active {
+    background-color: #ffdb58; /* Slightly different color on active click */
+}
+
+.nav-item a {
+    display: inline-block;
+}
+
+@media (max-width: 768px) {
+    .nav-bar {
+        flex-direction: column; /* Stack nav items on smaller screens */
+    }
+
+    .nav-item {
+        margin: 10px 0;
+    }
+}
+
+/* Dashboard Header */
+.dashboard-header {
+    text-align: center;
+    margin-bottom: 30px;
+}
+
+.dashboard-header h1 {
+    font-size: 2.5em;
+    color: #333;
+}
+
+.dashboard-header p {
+    font-size: 1.1em;
+    color: #556b7c; /* muted text */
+}
+
+/* Stat Cards */
+.stats-container {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 20px;
+    margin-bottom: 40px;
+}
+
+.stat-card {
+    background: white;
+    padding: 20px;
+    border-radius: 12px;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.1);
+    text-align: center;
+}
+
+.stat-card i {
+    font-size: 30px;
+    color: #4996d1;
+}
+
+.stat-card h3 {
+    margin-top: 10px;
+    font-size: 1.2em;
+}
+
+.stat-card p {
+    font-size: 1.5em;
+    font-weight: bold;
+    color: #4996d1;
+}
+
+/* Action Cards */
+.dashboard-actions {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+    justify-content: center;
+    margin: 30px 0;
+}
+
+.action-card {
+    background-color: white;
+    border-radius: 12px;
+    padding: 20px;
+    width: 200px;
+    text-align: center;
+    box-shadow: 0 2px 6px rgba(0,0,0,0.1);
+    color: #333;
+    text-decoration: none;
+    transition: background-color 0.3s;
+}
+
+.action-card i {
+    font-size: 28px;
+    color: #4996d1;
+    margin-bottom: 10px;
+}
+
+.action-card:hover {
+    background-color: #ffe87c;
+    color: #333;
+}
+
+/* Chart Container */
+.chart-container {
+    background: white;
+    padding: 20px;
+    border-radius: 12px;
+    box-shadow: 0 2px 6px rgba(0,0,0,0.1);
+    margin: 20px 0;
+}
+
+/* Buttons */
+button,
+.btn {
+    background-color: #4996d1;
+    color: white;
+    border: none;
+    padding: 10px 16px;
+    border-radius: 6px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+    font-size: 1em;
+}
+
+button:hover,
+.btn:hover {
+    background-color: #ffe87c;
+    color: #333;
+}
+
+/* Muted Text */
+.muted {
+    color: #556b7c;
+}
